Remove comparison with git-subrepo
This tool was originally inspired by git-subrepo, and git-subdir is
essentially aiming to be the tool I though git-subrepo was when I first
found it. I'm not sure what to put in the comparison since a number of
the issues I hit with git-subrepo seemed like bugs that (to me) made the
tool unusable and updating it to allow repeated rebases of local changes
on top of upstream changes (at least in the way I expected it to work)
required introducing the concept of integration branches and associating
upstream commit refs with squash commits which affects the overall
design. I suppose I never followed the usage model git-subrepo was
really geared for, considering that I wasn't able to use the tool myself
so I don't feel I can write a good comparison.